requirements.txt
pyyaml
requests
pyzabbix
